refactor(kv-transfer): move PD QoS injection into existing utils
