+++ vllm_ascend/distributed/kv_transfer/utils/utils.py
@@ -1,3 +1,4 @@
+import json
 import math
 import os
 from collections import OrderedDict, defaultdict
@@ -13,6 +14,25 @@ from vllm_ascend.distributed.parallel_state import get_p_tp_group
 
 MAX_HCCL_REGISTER_REGIONS = 256
 REGISTER_MERGE_GAP_BYTES = 4096
+
+QOS_KEY = "comm_resource_config.qos"
+PD_QOS_DEFAULT = 1
+QOS_MAX = 4
+
+
+def inject_qos(qos: int) -> None:
+    """Merge P/D QoS before engine initialization, preserving all other fields."""
+    if type(qos) is not int or not 0 <= qos <= QOS_MAX:
+        raise ValueError("kv_connector_extra_config.qos must be an integer in [0, 4]")
+    raw = os.getenv("ASCEND_GLOBAL_RESOURCE_CONFIG", "")
+    try:
+        config = json.loads(raw) if raw else {}
+    except (ValueError, TypeError) as exc:
+        raise ValueError("ASCEND_GLOBAL_RESOURCE_CONFIG must be a JSON object") from exc
+    if not isinstance(config, dict):
+        raise ValueError("ASCEND_GLOBAL_RESOURCE_CONFIG must be a JSON object")
+    config[QOS_KEY] = qos
+    os.environ["ASCEND_GLOBAL_RESOURCE_CONFIG"] = json.dumps(config)
 
 
 def kv_alltoall_and_rearrange(pd_tp_ratio: int, key: torch.Tensor, value: torch.TensorType):
